--- packages/client/src/api/changePassword.ts
@@ -1,11 +1,0 @@
-import { yandexApi } from "./setupApi"
-
-export type PasswordRequest = {
-    oldPassword: string,
-    newPassword: string  
-  }
-
-  export const putChangePassword = (request: PasswordRequest) => {
-    return yandexApi.put('user/password', request)
-    .catch((reason) => {console.log(reason)});
-  }
